MsbmsSystEnums.Types (msbms_syst_enums v0.1.0)

Types used by the Enums service module.

Types

enum_functional_type_name()

@type enum_functional_type_name() :: String.t()

Identification of each unique Enum Functional Type that can be associated with an Enum.

enum_functional_type_params()

@type enum_functional_type_params() :: %{
  optional(:enum_id) => Ecto.UUID.t(),
  optional(:internal_name) => enum_functional_type_name(),
  optional(:display_name) => String.t(),
  optional(:user_description) => String.t()
}

A map definition describing what specific key/value pairs are available for passing as SystEnumFunctionalTypes changeset parameters.

enum_item_id()

@type enum_item_id() :: binary()

Record ID type for the Enum Item record.

enum_item_name()

@type enum_item_name() :: String.t()

Identification of each unique Enum Value that can be associated with an Enum.

enum_item_params()

@type enum_item_params() :: %{
  optional(:enum_id) => Ecto.UUID.t(),
  optional(:functional_type_id) => Ecto.UUID.t(),
  optional(:functional_type_name) => enum_functional_type_name(),
  optional(:internal_name) => enum_item_name(),
  optional(:display_name) => String.t(),
  optional(:user_description) => String.t(),
  optional(:user_options) => map()
}

A map definition describing what specific key/value pairs are available for passing as SystEnumItems changeset parameters.

enum_name()

@type enum_name() :: String.t()

Identification of each unique Enum managed by the Enums Service instance.

enum_params()

@type enum_params() :: %{
  optional(:internal_name) => enum_name(),
  optional(:display_name) => String.t(),
  optional(:user_description) => String.t(),
  optional(:enum_items) => [enum_item_params()],
  optional(:functional_types) => [enum_functional_type_params()]
}

A map definition describing what specific key/value pairs are available for passing as SystEnums changeset parameters.

enum_service_params()

@type enum_service_params() ::
  {service_name(), MsbmsSystDatastore.Types.context_name()}

The expected form of the parameters used to start the Enums service.

service_name()

@type service_name() :: atom() | {:via, module(), atom() | {atom(), atom()}}

The valid forms of service name acceptable to identify the Enums service.

Currently we expect the service name to be an atom, though we expect that any of a simple local name, the :global registry, or the Registry module to be used for service registration. Any registry compatible with those options should also work.